Use correct typedef (int) for pointer-sized types

+++ userspace/lolibc/stddef.h
@@ -3,9 +3,9 @@
 
 #define NULL 0
 
-typedef long ssize_t;
-typedef unsigned long size_t;
-typedef long ptrdiff_t;
+typedef int ssize_t;
+typedef unsigned int size_t;
+typedef int ptrdiff_t;
 
 #define offsetof(type, member) \
     __builtin_offsetof(type, member)
